+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- Adds a Admin Dashboard for system wide analytics and management.
+
 ## [v19.0.3](https://github.com/lexicalunit/spellbot/releases/tag/v19.0.3) - 2026-05-21
 
 ### Changed